Moved interfaces into types
