+++ packages/keymaster/src/keymaster.ts
@@ -20,9 +20,11 @@ import {
     FileAssetOptions,
     CreateResponseOptions,
     EncryptOptions,
+    FileAsset,
     FixWalletResult,
     Group,
     IDInfo,
+    ImageAsset,
     IssueCredentialsOptions,
     KeymasterInterface,
     Poll,
@@ -70,21 +72,6 @@ export interface EncryptedMessage {
 
 interface PossiblySigned {
     signature?: Signature;
-}
-
-export interface Image {
-    cid: string;
-    type: string;
-    width: number;
-    height: number;
-    bytes: number;
-}
-
-export interface FileAsset {
-    cid: string;
-    type: string;
-    filename: string;
-    bytes: number;
 }
 
 export default class Keymaster implements KeymasterInterface {
@@ -738,9 +725,9 @@ export default class Keymaster implements KeymasterInterface {
         return this.updateAsset(id, data);
     }
 
-    async getImage(id: string): Promise<Image | null> {
+    async getImage(id: string): Promise<ImageAsset | null> {
         const asset = await this.resolveAsset(id);
-        const castAsset = asset as { image?: Image };
+        const castAsset = asset as { image?: ImageAsset };
 
         return castAsset.image ?? null;
     }
